+++ toughio/_io/output/tecplot/_tecplot.py
@@ -1,15 +1,30 @@
-from __future__ import with_statement
-
 import numpy as np
 
 from ...._common import open_file
-from ...._mesh.tecplot._tecplot import _read_variables, _read_zone
 from .._common import to_output
 
 __all__ = [
     "read",
     "write",
 ]
+
+
+zone_key_to_type = {
+    "T": str,
+    "I": int,
+    "J": int,
+    "K": int,
+    "N": int,
+    "NODES": int,
+    "E": int,
+    "ELEMENTS": int,
+    "F": str,
+    "ET": str,
+    "DATAPACKING": str,
+    "ZONETYPE": str,
+    "NV": int,
+    "VARLOCATION": str,
+}
 
 
 def read(filename, file_type, file_format, labels_order):
@@ -89,3 +104,80 @@ def write(filename, output):
             for d in data:
                 record = "".join("{:20.12e}".format(x) for x in d)
                 f.write("{}{}\n".format("{:18}".format(""), record))
+
+
+def _read_variables(line):
+    # Gather variables in a list
+    line = line.split("=")[1]
+    line = [x for x in line.replace(",", " ").split()]
+    variables = []
+
+    i = 0
+    while i < len(line):
+        if '"' in line[i] and not (line[i].startswith('"') and line[i].endswith('"')):
+            var = "{}_{}".format(line[i], line[i + 1])
+            i += 1
+        else:
+            var = line[i]
+
+        variables.append(var.replace('"', ""))
+        i += 1
+
+    # Check that at least X and Y are defined
+    if "X" not in variables and "x" not in variables:
+        raise ValueError("Variable 'X' not found")
+    if "Y" not in variables and "y" not in variables:
+        raise ValueError("Variable 'Y' not found")
+
+    return variables
+
+
+def _read_zone(line):
+    # Gather zone entries in a dict
+    line = line[5:]
+    zone = {}
+
+    # Look for zone title
+    ivar = line.find('"')
+
+    # If zone contains a title, process it and save the title
+    if ivar >= 0:
+        i1, i2 = ivar, ivar + line[ivar + 1 :].find('"') + 2
+        zone_title = line[i1 + 1 : i2 - 1]
+        line = line.replace(line[i1:i2], "PLACEHOLDER")
+    else:
+        zone_title = None
+
+    # Look for VARLOCATION (problematic since it contains both ',' and '=')
+    ivar = line.find("VARLOCATION")
+
+    # If zone contains VARLOCATION, process it and remove the key/value pair
+    if ivar >= 0:
+        i1, i2 = line.find("("), line.find(")")
+        zone["VARLOCATION"] = line[i1 : i2 + 1].replace(" ", "")
+        line = line[:ivar] + line[i2 + 1 :]
+
+    # Split remaining key/value pairs separated by '='
+    line = [x for x in line.replace(",", " ").split() if x != "="]
+    i = 0
+    while i < len(line) - 1:
+        if "=" in line[i]:
+            if not (line[i].startswith("=") or line[i].endswith("=")):
+                key, value = line[i].split("=")
+            else:
+                key = line[i].replace("=", "")
+                value = line[i + 1]
+                i += 1
+        else:
+            key = line[i]
+            value = line[i + 1].replace("=", "")
+            i += 1
+
+        zone[key] = zone_key_to_type[key](value)
+        i += 1
+
+    # Add zone title to zone dict
+    if zone_title:
+        zone["T"] = zone_title
+
+    return zone
