+++ toughio/_mesh/_common.py
@@ -1,6 +1,5 @@
 import logging
 
-import meshio
 import numpy as np
 
 vtk_to_meshio_type = {
@@ -83,99 +82,6 @@ meshio_type_to_ndim = {k: 3 for k in meshio_to_vtk_type.keys()}
 meshio_type_to_ndim.update(
     {"empty": 0, "vertex": 1, "line": 2, "triangle": 2, "polygon": 2, "quad": 2}
 )
-
-
-def get_meshio_version():
-    """
-    Return :mod:`meshio` version as a tuple.
-
-    Returns
-    -------
-    tuple
-        :mod:`meshio` version as tuple (major, minor, patch).
-
-    """
-    return tuple(int(i) for i in meshio.__version__.split("."))
-
-
-def get_old_meshio_cells(cells, cell_data=None):
-    """
-    Return old-style cells and cell_data (meshio < 4.0.0).
-
-    Parameters
-    ----------
-    cells : list of namedtuple (type, data)
-        New-style cells.
-    cell_data : dict or None, optional, default None
-        New-style cell data.
-
-    Returns
-    -------
-    dict
-        Old-style cells.
-    dict
-        Old-style cell data (only if `cell_data` is not None).
-
-    """
-    old_cells, cell_blocks = {}, {}
-    for ic, c in enumerate(cells):
-        if c.type not in old_cells.keys():
-            old_cells[c.type] = [c.data]
-            cell_blocks[c.type] = [ic]
-        else:
-            old_cells[c.type].append(c.data)
-            cell_blocks[c.type].append(ic)
-    old_cells = {k: np.concatenate(v) for k, v in old_cells.items()}
-
-    if cell_data is not None:
-        old_cell_data = (
-            {
-                cell_type: {
-                    k: np.concatenate([cell_data[k][i] for i in iblock])
-                    for k in cell_data.keys()
-                }
-                for cell_type, iblock in cell_blocks.items()
-            }
-            if cell_data
-            else {}
-        )
-        return old_cells, old_cell_data
-    else:
-        return old_cells
-
-
-def get_new_meshio_cells(cells, cell_data=None):
-    """
-    Return new-style cells and cell_data (meshio >= 4.0.0).
-
-    Parameters
-    ----------
-    cells : dict
-        Old-style cells.
-    cell_data : dict or None, optional, default None
-        Old-style cell data.
-
-    Returns
-    -------
-    list of namedtuple (type, data)
-        New-style cells.
-    dict
-        New-style cell data (only if `cell_data` is not None).
-
-    """
-    from ._mesh import CellBlock
-
-    new_cells = [CellBlock(k, v) for k, v in cells.items()]
-
-    if cell_data is not None:
-        labels = np.unique([kk for k, v in cell_data.items() for kk in v.keys()])
-        new_cell_data = {k: [] for k in labels}
-        for k in new_cell_data.keys():
-            for kk in cells.keys():
-                new_cell_data[k].append(cell_data[kk][k])
-        return new_cells, new_cell_data
-    else:
-        return new_cells
 
 
 def labeler(n_cells, label_length=None):
