--- toughio/_mesh/tecplot/__init__.py
@@ -1,7 +1,0 @@
-from .._helpers import register
-from ._tecplot import read, write
-
-__all__ = ["read", "write"]
-
-
-register("tecplot", [".dat", ".tec"], read, write)
